generate tests for DPG with generate-tests option
